Exception handling - delete

+++ src/main/java/com/github/kaykymatos/course/services/UserService.java
@@ -3,8 +3,11 @@ package com.github.kaykymatos.course.services;
 import com.github.kaykymatos.course.entities.User;
 import com.github.kaykymatos.course.repositories.UserRepository;
 import com.github.kaykymatos.course.resources.exceptions.ResourceExceptionHandler;
+import com.github.kaykymatos.course.services.exceptions.DatabaseException;
 import com.github.kaykymatos.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -30,7 +33,15 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        Optional<User> user = repository.findById(id);
+        if (user.isEmpty()) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User user) {
